+++ backend/accounting-service/k8.kts
@@ -3,7 +3,7 @@ package dk.sdu.cloud.k8
 
 bundle {
     name = "accounting"
-    version = "1.7.8"
+    version = "1.7.10"
 
     withAmbassador(null) {
         services.add(