+++ backend/app-store-service/k8.kts
@@ -3,7 +3,7 @@ package dk.sdu.cloud.k8
 
 bundle {
     name = "app-store"
-    version = "0.17.1"
+    version = "0.17.2"
 
     withAmbassador(null) {
         addSimpleMapping("/api/hpc")