+++ backend/kubernetes-monitor-service/k8.kts
@@ -3,7 +3,7 @@ package dk.sdu.cloud.k8
 
 bundle {
     name = "kubernetes-monitor"
-    version = "0.2.6"
+    version = "0.2.7"
 
     withAmbassador("/api/kubernetes/monitor") {}