+++ backend/notification-service/k8.kts
@@ -3,7 +3,7 @@ package dk.sdu.cloud.k8
 
 bundle {
     name = "notification"
-    version = "1.5.7"
+    version = "1.5.8"
 
     withAmbassador("/api/notifications") {}