+++ backend/project-service/k8.kts
@@ -3,7 +3,7 @@ package dk.sdu.cloud.k8
 
 bundle { ctx ->
     name = "project"
-    version = "3.5.7"
+    version = "3.5.8"
 
     withAmbassador("/api/projects") {}