+++ wiki/release-notes.md
@@ -7,6 +7,46 @@ use.
 Note that each service may have more than one deployment associated with it.
 For details look in the `k8.kts` folders of each microservice.
 
+## 16-12-21
+
+Upgrade log4j to version 2.16.0.
+
+```
+✅  Deployment(accounting, 1.7.10) (UP-TO-DATE)
+✅  Deployment(password-reset, 0.4.8) (UP-TO-DATE)
+✅  Deployment(file-stats, 2.4.7) (UP-TO-DATE)
+✅  Deployment(mail, 0.4.10) (UP-TO-DATE)
+✅  Deployment(sync-mounter, 0.1.8) (UP-TO-DATE)
+✅  Deployment(alerting, 1.4.9) (UP-TO-DATE)
+✅  Deployment(activity, 1.7.9) (UP-TO-DATE)
+✅  Deployment(contact-book, 0.4.7) (UP-TO-DATE)
+✅  Deployment(share, 1.10.8) (UP-TO-DATE)
+✅  Deployment(slack, 0.3.7) (UP-TO-DATE)
+✅  Deployment(auth, 1.32.7) (UP-TO-DATE)
+✅  Deployment(support, 1.6.7) (UP-TO-DATE)
+✅  Deployment(task, 0.5.7) (UP-TO-DATE)
+✅  Deployment(notification, 1.5.8) (UP-TO-DATE)
+✅  Deployment(webdav, 0.4.8) (UP-TO-DATE)
+✅  Deployment(provider, 0.2.7) (UP-TO-DATE)
+✅  Deployment(audit-ingestion, 0.4.7) (UP-TO-DATE)
+✅  Deployment(storage, 4.5.4) (UP-TO-DATE)
+✅  Deployment(grant, 0.4.7) (UP-TO-DATE)
+✅  Deployment(ucloud-data-extraction, 0.1.18) (UP-TO-DATE)
+✅  Deployment(app-kubernetes, 0.22.2) (UP-TO-DATE)
+✅  Deployment(file-favorite, 1.8.7) (UP-TO-DATE)
+✅  Deployment(news, 0.3.7) (UP-TO-DATE)
+✅  Deployment(app-aau, 0.2.11) (UP-TO-DATE)
+✅  Deployment(project, 3.5.8) (UP-TO-DATE)
+✅  Deployment(filesearch, 1.5.7) (UP-TO-DATE)
+✅  Deployment(project-repository, 0.4.7) (UP-TO-DATE)
+✅  Deployment(app-store, 0.17.2) (UP-TO-DATE)
+✅  Deployment(app-orchestrator, 2.7.2) (UP-TO-DATE)
+✅  Deployment(avatar, 1.6.7) (UP-TO-DATE)
+✅  Deployment(kubernetes-monitor, 0.2.7) (UP-TO-DATE)
+✅  Deployment(indexing, 1.19.17) (UP-TO-DATE)
+✅  Deployment(file-trash, 1.7.7) (UP-TO-DATE)
+```
+
 ## 13-12-21
 
 Bug fix for upcoming downtime icon.